Following changes after ser/repl of size_t with CuSize_t. Many functions returned only void. They now return CuError_t. Added several NULL ptr checks. Added cunit target to project

+++ CuTestAssertArrayEqualsLineMsg.c
@@ -1,7 +1,7 @@
 
 #include "CuTest_internal.h"
 
-bool CuAssertArrayEquals_LineMsg(CuTest* tc, const char* file, unsigned int line, const char* message,
+bool CuAssertArrayEquals_LineMsg(CuTest* tc, const char* file, unsigned long int line, const char* message,
 	const void* expected, const void* actual, CuSize_t elementsize, CuSize_t len)
 {
     CuSize_t i=0,j=0;
